# pathocert-frontend/documentation/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: PathoTHREAT Backend
  description: The API available by the PathoTHREAT frontend to implement all the
    functionalities of the PathoTHREAT tool.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://localhost:8080
  description: PathoTHREAT Backend Localhost Server
- url: https://localhost:4200/api
  description: PathoTHREAT Backend Localhost Server With Frontend
...
paths:
...
  /api/expert/search-terms:
    get:
      tags:
      - Expert Asking
      summary: Get Expert Search Result
      description: "Search for information in the expert database. The endpoint receives\
        \ a subject and a desired output, and returns the values found."
      operationId: api_expert_search_terms_get
      parameters:
      - name: token
        in: header
        description: Token of the authenticated user.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/AuthenticateHeader'
      - name: subject
        in: query
        description: Subject of the query.
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: earthquake
      - name: desiredOutput
        in: query
        description: Type of the desired outputs.
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: Contaminant
      responses:
        "200":
          description: The search was executed correctly.
          content:
            application/hal+json:
              schema:
                type: string
                description: String in HTML format to be shown directly to the user.
                example: <p>ecoli<br>norovirus<br></p>
                x-content-type: application/hal+json
        "5XX":
          description: Something went wrong while executing the query.
      x-openapi-router-controller: swagger_server.controllers.expert_asking_controller
  /api/expert/values:
    get:
      tags:
      - Expert Asking
      summary: Get All Values In Expert Data
      description: Get a list of all the values found in the expert database. This
        is used to provide autocomplete functionality to the search component.
      operationId: api_expert_values_get
      parameters:
      - name: token
        in: header
        description: Token of the authenticated user.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/AuthenticateHeader'
      responses:
        "200":
          description: The query was executed correctly.
          content:
            application/hal+json:
              schema:
                type: array
                example:
                - ecoli
                - norovirus
                - earthquake
                - '...'
                items:
                  type: string
                x-content-type: application/hal+json
        "5XX":
          description: Something went wrong while executing the query.
      x-openapi-router-controller: swagger_server.controllers.expert_asking_controller
  /api/expert/labels:
    get:
      tags:
      - Expert Asking
      summary: Get All Labels In Expert Data
      description: Get all the labels found in the expert data. This is used to provide
        autocomplete functionality to the search component.
      operationId: api_expert_labels_get
      parameters:
      - name: token
        in: header
        description: Token of the authenticated user.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/AuthenticateHeader'
      responses:
        "200":
          description: The endpoint was executed correctly.
          content:
            application/hal+json:
              schema:
                type: array
                example:
                - Event
                - CascadingWNEvent
                - ProducedEvent
                - Contaminant
                - EffectWaterTaste
                - EffectWaterOdor
                - EffectWater
                - ContaminantFamily
                - ContaminantTreatment
                - EffectHealth
                - Symptom
                - ContaminantType
                items:
                  type: string
                x-content-type: application/hal+json
        "5XX":
          description: Something went wrong while executing the query.
      x-openapi-router-controller: swagger_server.controllers.expert_asking_controller
  /api/expert/values/{type}:
    get:
      tags:
      - Expert Asking
      summary: Get All Values Of Certain Type In Expert Data
      description: Get a list of all the values of a certain type found in the expert
        database. This is used to provide extra functionalities to the autocomplete
        functionality in the search component.
      operationId: api_expert_values_type_get
      parameters:
      - name: token
        in: header
        description: Token of the authenticated user.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/AuthenticateHeader'
      - name: type
        in: path
        description: Type of the values the user is interested in.
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Symptom
      responses:
        "200":
          description: The endpoint was executed correctly.
          content:
            application/hal+json:
              schema:
                type: array
                example:
                - diarrhea
                - vomiting
                items:
                  type: string
                x-content-type: application/hal+json
        "5XX":
          description: Something went wrong while executing the query.
      x-openapi-router-controller: swagger_server.controllers.expert_asking_controller
  /api/expert/labels/{subject}:
    get:
      tags:
      - Expert Asking
      summary: Get All Labels Connected To Subject In Expert Data
      description: Get a list of all the labels connected to the specified subject.
        This is used to provide extra functionalities to the autocomplete functionality
        in the search component.
      operationId: api_expert_labels_subject_get
      parameters:
      - name: token
        in: header
        description: Token of the authenticated user.
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/AuthenticateHeader'
      - name: subject
        in: path
        description: Subject for which we want to check related labels.
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: ecoli
      responses:
        "200":
          description: The endpoint was executed correctly.
          content:
            application/hal+json:
              schema:
                type: array
                example:
                - Event
                - ProducedEvent
                - Symptom
                - EffectHealth
                items:
                  type: string
                x-content-type: application/hal+json
        "5XX":
          description: Something went wrong while executing the query.
      x-openapi-router-controller: swagger_server.controllers.expert_asking_controller
components:
  schemas:
...
    AuthenticateHeader:
      type: string
      description: Example of the value inside an Authorization header in most requests
        in order to authenticate the user.
      example: Bearer eyJhbGciOiJSUzI1...
...